fixed bug in reconnect & disconnect
@ws.close does not actually always work in those methods, because @ws is nil when the stream is currently disconnected but trying to reconnect. So this would throw "undefined method `close' for nil:NilClass" if you tried to disconnect or reconnect while it was already reconnecting.

So instead, I'm rolling back to the previous approach of EM.stop_event_loop to force a disconnect & shutdown in any scenario, and in #reconnect we need to check if there is a @ws, and if not, we bypass the onclose block and call #connect directly.

(I'm not sure why I actually changed EM.stop_event_loop to @ws.close previously... I think my thinking was that it's better to "cleanly" shut down by passing through the onclose block. This would also ensure that the subscriber's on disconnect callback is called then. So right now if you call #disconnect explicitly while the stream is reconnecting, it will not call your on disconnect callback. Is this correct or incorrect behavior? ¯\_(ツ)_/¯ tbh)