don't start heartbeat by default
There are some firehoses which will not be sending events often, possibly only once in a while - e.g. labellers and independent PDS firehoses - and in this case we don't want any heartbeat since it will be normal not to have any events for a long time. It's not really possible to detect easily if we're connecting to a full network relay or one of those, so it's best not to enable this unless it's enabled manually.